// src/types/ORCType.ts
export type ORCAllowances = {
  WindSpeeds: number[];
  WindAngles: number[];
  R52: number[];
  R60: number[];
  R75: number[];
  R90: number[];
  R110: number[];
  R120: number[];
  R135: number[];
  R150: number[];
  Beat: number[];
  Run: number[];
  BeatAngle: number[];
  GybeAngle: number[];
  WL: number[];
  CR: number[];
  OC: number[];
};

export type RMS = {
  NatAuth: string;
  BIN: string;
  CertNo: string;
  RefNo: string;
  SailNo: string;
  YachtName: string;
  Class: string;
  Builder: string;
  Designer: string;
  Address3: string;
  C_Type: string;
  Family: string;
  Division: string;
  IssueDate: string;
  Dspl_Sailing: number;
  WSS: number;
  Area_Main: number;
  Area_Jib: number;
  Area_Sym: number;
  Area_Asym: number;
  Age_Year: number;
  CrewWT: number;
  LOA: number;
  IMSL: number;
  Draft: number;
  MB: number;
  Dspl_Measurement: number;
  Stability_Index: number;
  Dynamic_Allowance: number;
  GPH: number;
  CDL: number;
  ILCWA: number;
  TMF_Inshore: number;
  APHD: number;
  APHT: number;
  OSN: number;
  TMF_Offshore: number;
  TND_Offshore_Low: number;
  TN_Offshore_Low: number;
  TND_Offshore_Medium: number;
  TN_Offshore_Medium: number;
  TND_Offshore_High: number;
  TN_Offshore_High: number;
  TND_Inshore_Low: number;
  TN_Inshore_Low: number;
  TND_Inshore_Medium: number;
  TN_Inshore_Medium: number;
  TND_Inshore_High: number;
  TN_Inshore_High: number;
  Pred_Up_TOD: number;
  Pred_Up_TOT: number;
  Pred_Down_TOD: number;
  Pred_Down_TOT: number;
  US_PREDUP_L_TOD: number;
  US_PREDUP_L_TOT: number;
  US_PREDUP_M_TOD: number;
  US_PREDUP_M_TOT: number;
  US_PREDUP_H_TOD: number;
  US_PREDUP_H_TOT: number;
  US_PREDDN_L_TOD: number;
  US_PREDDN_L_TOT: number;
  US_PREDDN_M_TOD: number;
  US_PREDDN_M_TOT: number;
  US_PREDDN_H_TOD: number;
  US_PREDDN_H_TOT: number;
  US_CHIMAC_UP_TOT: number;
  US_CHIMAC_AP_TOT: number;
  US_CHIMAC_DN_TOT: number;
  US_BAYMAC_CV_TOT: number;
  US_BAYMAC_SH_TOT: number;
  US_HARVMOON_TOD: number;
  US_HARVMOON_TOT: number;
  US_VICMAUI_TOT: number;
  KR_PREDR_TOD: number;
  RSA_CD_INS_TOD: number;
  RSA_CD_INS_TOT: number;
  RSA_CD_OFF_TOD: number;
  RSA_CD_OFF_TOT: number;
  Allowances: ORCAllowances;
};